Moved 'Main' to 'Core'.
Updated READMEs, first pass.

Updated Cabal file.

Updated Tests to handle change of Main to Core.
