+++ tests/Level06Tests.hs
@@ -12,8 +12,8 @@ import           Test.Hspec.Wai
 import qualified System.Exit    as Exit
 
 import qualified Level06.AppM   as AppM
+import qualified Level06.Core   as Core
 import qualified Level06.DB     as DB
-import qualified Level06.Main   as Main
 import qualified Level06.Types  as Types
 
 doctests :: [FilePath]
@@ -26,23 +26,22 @@ unitTests :: IO ()
 unitTests = do
   let dieWith m = print m >> Exit.exitFailure
 
-  reqsE <- Main.prepareAppReqs
+  reqsE <- Core.prepareAppReqs
   case reqsE of
 
     Left err -> dieWith err
 
     Right ( cfg, db ) -> do
-      let app' = pure ( Main.app cfg db )
+      let app' = pure (Core.app cfg db)
 
           flushTopic =
             -- Clean up and yell about our errors
-            (either dieWith pure =<<) $ AppM.runAppM
-            ( do
-                -- We don't export the constructor so even for known values we have
-                -- to play by the rules. There is no - "Oh just this one time.", do it right.
-                t <- AppM.liftEither $ Types.mkTopic "fudge"
-                -- Purge all of the comments for this topic for our tests
-                DB.deleteTopic db t
+            either dieWith pure =<< AppM.runAppM (
+            -- We don't export the constructor so even for known values we have
+            -- to play by the rules. There is no - "Oh just this one time.", do it right.
+            AppM.liftEither (Types.mkTopic "fudge")
+              -- Purge all of the comments for this topic for our tests
+              >>= DB.deleteTopic db
             )
 
       -- Run the tests with a DB topic flush between each spec
